Fix NullReferenceException on Image
Every once in a while, Image will have a value so it schedules code on
the dispatcher, but then Image is set to null and finally the dispatched
method gets called. So I added an additional check for null inside the
dispatched action. I also fixed the possibility of Image becoming null
after the null check, but before calling the dispatcher (rare, but
possible).